fix: fixed build errors
