+++ clients/giggle-client/src/components/SignUpForm/SignUpForm.tsx
@@ -15,7 +15,7 @@ import SignUpProcessPassword from '../SignUpProcess/SignUpProcessPassword'
 import SignUpProcessUserInfo from '../SignUpProcess/SignUpProcessUserInfo'
 import SignUpProcessTermsAndConditions from '../SignUpProcess/SignUpProcessTermsAndConditions'
 import { ResultReturnType } from '@/libs/types'
-import { API_AUTH_GET_GOOGLE_ERROR_CODE } from '@/app/api/auth/google/route'
+import { API_AUTH_GET_GOOGLE_ERROR_CODE } from '@/app/api/auth/google/types'
 
 const TITLE_MESSAGE = `Sign up to start finding venues`
